100 no mooshak com enum

+++ src/Package/Services/Eating.java
@@ -2,7 +2,7 @@ package Package.Services;
 
 public class Eating extends ServicesWithCapacity {
 
-    public Eating(long latitude, long longitude, int price, int capacity,String name){
+    public Eating(long latitude, long longitude, float price, int capacity,String name){
         super(latitude,longitude,price,capacity,name,"eating");
     }
 }
